requirements.txt
alembic==1.7.7
amqp==5.1.1
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.2
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.1
flower==1.0.0
GeoAlchemy2==0.11.1
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.2
h11==0.13.0
httptools==0.2.0
humanize==4.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.1
kombu==5.2.4
Mako==1.2.0
MarkupSafe==2.1.1
orjson==3.6.7
packaging==21.3
passlib==1.7.4
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.9
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==5.4.1
redis==4.3.2
requests==2.27.1
rsa==4.8
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
tornado==6.1
typing_extensions==4.1.1
fastapi_utils==0.2.1
ujson==4.3.0
urllib3==1.26.9
uvicorn==0.15.0
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.2
wrapt==1.14.1
zipp==3.8.0
